refactor: change event handler naming convention
