+++ server/events/onMapNameAssigned.spec.ts
