+++ server/events/onMapNameAssigned.ts
