+++ server/events/onTokenAssigned.spec.ts
