+++ server/events/onTokenAssigned.ts
